Fixed urls for Mezuro
Without this, Colab was unable to request assets or any other sub page from
Mezuro.

This is not the correct fix. The right place to set the url is on:
colab/plugins/mezuro/urls.py. But this is not working yet.

Notice that we've also identified a bug on the ReverseProxy, which is not
repecting the upstream port always trying to request from port 80.

Signed off by: Gustavo Jaruga <darksshades@hotmail.com>